Added link depth and hidden links in graph
Added a link depth field to the database to track the minimum number of
links that need to be followed to get to a url. Links that are common
across multiple pages are now hidden to stop the graph becoming a huge
mess.

Reworked how page titles are collected. ( still needs some work).